requirements.txt
mysql-connector-python
python-dotenv
python-telegram-bot
flask
beautifulsoup4
requests
lxml


